requirements.txt
PyQt6>=6.6.0
psutil>=5.9.0
pynvml>=11.5.0
liquidctl>=1.13.0
openrgb-python>=0.2.13
pyinstaller>=6.0.0
Pillow>=10.0.0
pywin32>=306; sys_platform == "win32"
WMI>=1.5.1; sys_platform == "win32"
